requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
SQLAlchemy>=2.0
pydantic>=2.5
pydantic[email]>=2.5
python-dotenv>=1.0
dateparser>=1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psycopg2>=2.9.0
gunicorn>=21.0.0